+++ lglib.php
@@ -1,15 +1,15 @@
 <?php
 /**
-*   Table names and other global configuration values.
-*
-*   @author     Lee Garner <lee@leegarner.com>
-*   @copyright  Copyright (c) 2012-2020 Lee Garner <lee@leegarner.com>
-*   @package    lglib
-*   @version    1.0.12
-*   @license    http://opensource.org/licenses/gpl-2.0.php 
-*               GNU Public License v2 or later
-*   @filesource
-*/
+ * Table names and other global configuration values.
+ *
+ * @author      Lee Garner <lee@leegarner.com>
+ * @copyright   Copyright (c) 2012-2021 Lee Garner <lee@leegarner.com>
+ * @package     lglib
+ * @version     v1.0.12
+ * @license     http://opensource.org/licenses/gpl-2.0.php 
+ *              GNU Public License v2 or later
+ * @filesource
+ */
 
 // Static configuration items
 global $_LGLIB_CONF;
@@ -23,5 +23,3 @@ global $_TABLES, $_DB_table_prefix;
 $_DB_prefix = $_DB_table_prefix . 'lglib_';
 $_TABLES['lglib_messages'] = $_DB_prefix . 'messages';
 $_TABLES['lglib_jobqueue'] = $_DB_prefix . 'jobqueue';
-
-?>