+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## 1.0.12
 Release TBD
+  * SmartResizer: do not create thumbnail directory until needed.
   * Move Message and JobQueue functions into classes.
   * Enable SmartResizer to return full image urls and skip lightbox tags.
   * Lower PHP requirement to 7.2.0.
   * Allow plugins to register templates to be used by SmartResizer.
   * Standardize on `nosmartresize` to avoid smart resizing.
+  * Add `success` as a message level.
 
 ## 1.0.11
 Release 2020-07-27